Switch Case and getting mic to work
created switch case to determine note based on the buttons being pressed, there are issues since 16 notes use a pinch off of the back button and we don't have that implimented. I also had alot of issues trying to get the mediaplayer to play when the blowing threshold is met and stop when the blowing stops. Will talk about more in lab Wed.